// generated/src/app/api-models/prov-address.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProvAddress {
  seqProvAddress: number;
  seqProvId: number;
  name2: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  county: string;
  medicalGroup: string;
  sundayHours: string;
  mondayHours: string;
  tuesdayHours: string;
  wednesdayHours: string;
  thursdayHours: string;
  fridayHours: string;
  saturdayHours: string;
  userDefined1: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  primaryAddress: string;
  country: string;
  longitude: number;
  latitude: number;
  geoResult: string;
  userDefined2: string;
  userDate1: Date;
  userDate2: Date;
  siteCode: string;
  seqSiteAddressId: number;
}